+++ libraries/ArduinoPin/test/ArduinoPin_test.cpp
@@ -2,8 +2,8 @@
 
 #include "MockArduinoPin.h"
 
-#define digitalPinToTimer(x) (((x) == 2) ? 9 : NOT_ON_TIMER)
-#define digitalPinToInterrupt(x) (((x) == 5) ? 3 : NOT_AN_INTERRUPT)
+#define digitalPinToTimer(x) (uint8_t(((x) == 2) ? 9 : NOT_ON_TIMER))
+#define digitalPinToInterrupt(x) (uint8_t(((x) == 5) ? 3 : NOT_AN_INTERRUPT))
 
 static void voidFunc(void) {
 }
